feat: Add overrideable config and overall improvement to train command
- Moved and renamed module files around for better organization
