+++ Cargo.toml
@@ -1,3 +1,4 @@
+
 [workspace]
 # Try
 # require version 2 to avoid "feature" additiveness for dev-dependencies
@@ -44,7 +45,7 @@ uuid = { version = "1.17.0", features = [
     "serde",
 ] }
 crossbeam = { version = "0.8.4", default-features = false, features = ["crossbeam-channel"]}
-
+json-patch = { version = "*", default-features = false }
 
 ### For xtask crate ###
 tracel-xtask = { version = "=1.1.8" }
