--- crates/burn-central-cli/src/cli_commands/mod.rs
@@ -1,5 +1,0 @@
-pub mod run;
-
-pub mod init;
-pub mod login;
-pub mod package;
