+++ crates/burn-central-cli/src/commands/init.rs
