+++ crates/burn-central-cli/src/lib.rs
@@ -2,8 +2,8 @@ pub mod cli;
 pub mod config;
 pub mod registry;
 
-mod cli_commands;
 mod commands;
+mod execution;
 mod logging;
 
 mod context;
