+++ crates/burn-central-cli/src/registry.rs
@@ -33,6 +33,8 @@ pub const fn make_static_lazy<T: 'static>(func: fn() -> T) -> LazyValue<T> {
 
 pub use inventory;
 pub use paste;
+use quote::ToTokens;
+use burn_central_client::schemas::RegisteredFunction;
 
 // macro that generates a flag with a given type and arbitrary parameters and submits it to the inventory
 #[macro_export]
@@ -52,5 +54,25 @@ pub(crate) fn get_flags() -> Vec<Flag> {
     inventory::iter::<Plugin<Flag>>
         .into_iter()
         .map(|plugin| (*plugin.0).to_owned())
+        .collect()
+}
+
+pub fn get_registered_functions(flags: &[Flag]) -> Vec<RegisteredFunction> {
+    flags
+        .iter()
+        .map(|flag| {
+            // function token stream to readable string
+            let itemfn = syn_serde::json::from_slice::<syn::ItemFn>(flag.token_stream)
+                .expect("Should be able to parse token stream.");
+            let syn_tree: syn::File = syn::parse2(itemfn.into_token_stream())
+                .expect("Should be able to parse token stream.");
+            let code_str = prettyplease::unparse(&syn_tree);
+            RegisteredFunction {
+                mod_path: flag.mod_path.to_string(),
+                fn_name: flag.fn_name.to_string(),
+                proc_type: flag.proc_type.to_string(),
+                code: code_str,
+            }
+        })
         .collect()
 }
