+++ crates/burn-central-cli/src/util/mod.rs
@@ -1,2 +1,3 @@
 pub mod cargo;
 pub mod git;
+pub mod time;
