+++ crates/burn-central-client/src/experiment/mod.rs
@@ -1,8 +1,10 @@
 mod base;
+mod config;
 mod error;
 mod log_store;
 mod message;
 mod socket;
 
 pub use base::*;
+pub use config::*;
 pub use error::*;
